unspecified-css-subresources/import.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>@import blocks load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
</script>
<style>
@import "import.css?pipe=trickle(d5)";
</style>
<div id=foo>XXX</div>
<script>
var t1 = performance.timing.responseStart;

async_test(function(t) {
  window.addEventListener("load", t.step_func_done(function() {
    var t2 = Date.now();
    var diff = t2 - t1;
    assert_greater_than(diff, 5000, "ms after style");
    var style = getComputedStyle(document.querySelector("#foo"));
    assert_equals(style.getPropertyValue("color"), "rgb(0, 0, 255)", "#foo.color");
  }));
}, "load waits for @import");
</script>
